Corrected a few features.
Fixed the autoplay when a song ends.
Improved the update method to only update when the player is supposed to
be playing songs.
Fixed audio files not loading properly.
Fixed invalid audio files loading.
Changed the randomness toggling to keep track of where we were in the
ordered list of songs.